Added Reference to Markdig in the xml doc
I needed to look up how to write Abbreviation in MarkDown. The documentation stated how you can enable them, but not how to use them.

I think it would be a good Idea to link to the implementation for questions how to use MarkDown with the specific features of the implementation.